+++ src/cli/chain/index.js
@@ -1,1 +1,3 @@
 module.exports = require('./create');
+module.exports = require('./list');
+module.exports = require('./remove');