+++ src/cli/chain/remove/index.js
@@ -1,0 +1,26 @@
+const chalk = require('chalk');
+
+const { cli, action } = require('core/cli');
+const { print } = require('core/util');
+
+const { execute, run } = require('./remove');
+
+cli('chain:remove [<chainName>]', 'Remove chains', input => action(execute, run, input), {
+  alias: 'reset',
+  requirements: {
+    forgeRelease: true,
+    runningNode: false,
+    rpcClient: false,
+    wallet: false,
+  },
+  options: [['-y, --yes', 'Use default value for all options']],
+  handlers: {
+    '--help': () => {
+      print(`
+Examples:
+- ${chalk.cyan('forge chain:remove chain-1')}           remove interactively
+- ${chalk.cyan('forge chain:remove chain-1 --yes')}     force remove
+      `);
+    },
+  },
+});