+++ src/core/forge-process.js
@@ -6,30 +6,10 @@ const findProcess = require('find-process');
 const shell = require('shelljs');
 
 const debug = require('./debug')('forge-process');
-const { getAllAppNames } = require('./forge-config');
-const { getTendermintHomeDir } = require('./forge-fs');
-const { prettyTime, md5, sleep } = require('./util');
-const { DEFAULT_CHAIN_NAME } = require('../constant');
+const { getTendermintHomeDir, getAllChainNames } = require('./forge-fs');
+const { prettyTime, md5, sleep, chainSortHandler } = require('./util');
 
-const processSortHandler = (x, y) => {
-  // make default chain the first in order
-  if (x.name === DEFAULT_CHAIN_NAME) {
-    return -1;
-  }
-
-  if (y.name === DEFAULT_CHAIN_NAME) {
-    return 1;
-  }
-
-  if (x.name > y.name) {
-    return 1;
-  }
-  if (x.name < y.name) {
-    return -1;
-  }
-
-  return 0;
-};
+const sortHandler = (x, y) => chainSortHandler(x.name, y.name);
 
 const getProcessTag = (name, chainName = process.env.FORGE_CURRENT_CHAIN) => {
   if (!name) {
@@ -45,8 +25,8 @@ async function findServicePid(n) {
   return match ? match.pid : 0;
 }
 
-async function getTendermintProcess(appName) {
-  const homeDir = getTendermintHomeDir(appName);
+async function getTendermintProcess(chainName) {
+  const homeDir = getTendermintHomeDir(chainName);
   const tendermintProcess = await findProcess('name', 'tendermint');
 
   const tmp = tendermintProcess.find(({ cmd }) => cmd.includes(homeDir));
@@ -60,8 +40,8 @@ async function findForgeEpmdDeamon() {
   return { name: 'forge-epmd', pid: tmp ? tmp.pid : 0 };
 }
 
-async function isForgeStarted(appName = process.env.FORGE_CURRENT_CHAIN) {
-  const { pid } = await getTendermintProcess(appName);
+async function isForgeStarted(chainName = process.env.FORGE_CURRENT_CHAIN) {
+  const { pid } = await getTendermintProcess(chainName);
 
   return !!pid;
 }
@@ -136,34 +116,34 @@ async function getRunningProcessesStats(chainName = process.env.FORGE_CURRENT_CH
 
 async function getAllRunningProcessStats() {
   const processes = [];
-  const allAppNames = getAllAppNames();
+  const allChains = getAllChainNames();
 
   // eslint-disable-next-line
-  for (const appName of allAppNames) {
+  for (const name of allChains) {
     // eslint-disable-next-line
-    const tmp = await getRunningProcessesStats(appName);
+    const tmp = await getRunningProcessesStats(name);
     if (tmp && tmp.length) {
-      processes.push({ name: appName, value: tmp });
+      processes.push({ name, value: tmp });
     }
   }
 
-  return processes.sort(processSortHandler);
+  return processes.sort(sortHandler);
 }
 
 async function getAllProcesses() {
   const processes = [];
-  const allAppNames = getAllAppNames();
+  const allChains = getAllChainNames();
 
   // eslint-disable-next-line
-  for (const appName of allAppNames) {
+  for (const name of allChains) {
     // eslint-disable-next-line
-    const tmp = await getRunningProcesses(appName);
+    const tmp = await getRunningProcesses(name);
     if (tmp && tmp.length) {
-      processes.push({ name: appName, value: tmp });
+      processes.push({ name, value: tmp });
     }
   }
 
-  processes.sort(processSortHandler);
+  processes.sort(sortHandler);
 
   return processes;
 }
@@ -223,12 +203,12 @@ async function stopForgeProcesses(chainName) {
 }
 
 module.exports = {
+  findServicePid,
+  isForgeStarted,
   getAllProcesses,
   getAllRunningProcesses,
   getAllRunningProcessStats,
   getRunningProcesses,
-  findServicePid,
-  isForgeStarted,
   getForgeProcess,
   getForgeWebProcess,
   getForgeWorkshopProcess,