+++ tools/update-readme.js
@@ -4,7 +4,16 @@ const path = require('path');
 const shell = require('shelljs');
 
 const readmeFile = path.join(__dirname, '../README.md');
-const { stdout } = shell.exec(path.resolve(__dirname, '../bin/forge'), { silent: true });
+const { code, stdout, stderr } = shell.exec(path.resolve(__dirname, '../bin/forge'), {
+  silent: true,
+});
+
+if (code > 0) {
+  console.error(`get forge help error ${code}:`);
+  console.error(stderr);
+  process.exit(code);
+}
+
 const readmeContent = `![forge-cli](https://www.arcblock.io/.netlify/functions/badge/?text=Forge%20CLI)
 
 > Command line toolbox maintained by [Arcblock](https://www.arcblock.io) that helps developers to work with [Forge SDK](https://docs.arcblock.io/forge/latest/)
@@ -20,11 +29,11 @@ const readmeContent = `![forge-cli](https://www.arcblock.io/.netlify/functions/b
 
 ## Introduction
 
-[![](https://img.shields.io/npm/v/@arcblock/forge-cli.svg?label=forge-cli)](https://www.npmjs.com/package/@arcblock/forge-cli)
-[![docs](https://img.shields.io/badge/powered%20by-arcblock-green.svg)](https://docs.arcblock.io/forge/latest/tools/forge_cli.html)
-[![Build Status](https://img.shields.io/travis/arcblock/forge-cli.svg?style=flat-square)](https://travis-ci.com/arcblock/forge-cli)
+[![](https://img.shields.io/npm/v/@arcblock/forge-cli.svg?label=forge-cli&style=flat-square)](https://www.npmjs.com/package/@arcblock/forge-cli)
+[![docs](https://img.shields.io/badge/powered%20by-arcblock-green.svg?style=flat-square)](https://docs.arcblock.io/forge/latest/tools/forge_cli.html)
+[![Build Status](https://img.shields.io/travis/com/arcblock/forge-cli?style=flat-square)](https://travis-ci.com/arcblock/forge-cli)
 [![PRs Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg?style=flat-square)](http://makeapullrequest.com)
-[![Gitter](https://badges.gitter.im/ArcBlock/community.svg)](https://gitter.im/ArcBlock/community?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge)
+[![Gitter](https://img.shields.io/gitter/room/ArcBlock/community?style=flat-square&color=%234cb696)](https://gitter.im/ArcBlock/community?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge)
 
 \`forge-cli\` is an awesome toolbox for developers to work with forge, and forge is an awesome framework for building decentralized applications. Out of the box forge-cli support following features: